fix: istio integration
Removes the deprecated service-mesh interface and instead drives istio integration based on whether an ingress is provided.

This may not be sufficient as Istio integration may be useful even when an ingress is not provided.  If so, we can add a second config option to enable Istio separately from the ingress.