tests: remove SeldonDeployments even if the test case fails
Using a "remove" fixture we can guarantee that the SeldonDeployment
under test is actually removed from the testing environment before
proceeding with the test execution. This change prevents potential
conflicts when the SeldonDeployment has the same name as a previously
deployed one and keeps the testing environment clean before continuing.

This commit also moves some of the environment preparations to pytest
fixtures to keep the test cases tidy and the test file reusable.

Part of #185