tests(integration): Apply seldondeployments in namespace `default` (#220)
* tests(integration): Refactor

Refactor integration tests after hitting issue #218. This is a
workaround for the issue. At the same time, it is needed to revert those
back since up until now, integration tests were applying
seldondeployments to the namespace created by Juju. In production, this
would be `kubeflow` and we would never apply something there.

* tests(integration): Rename DEFAULT_NAMESPACE to WORKLOADS_NAMESPACE