Add logging relation to seldon-core (#259)
* Add logging relation to seldon-core

Use LogForwarder from loki_push_api library to forwards all logs from
workload container to Loki. This requires to use ops 2.9 or newer, so it
was needed to be updated.

fixes: #253

* bump all integration requirements

We are testing on Juju 3.5, but we use libjuju 3.2.2. This version do
not support websockets 12.0, however on my system I reqire to have
at-least 10.0.